+++ commons/src/main/java/commons/Question.java
@@ -3,5 +3,5 @@ package commons;
 public abstract class Question {
 	public Question() {}
 
-	public abstract int pointsEarned(int maxPoints, int answerGiven, double progress);
+	public abstract int pointsEarned(int maxPoints, long answerGiven, double progress);
 }
